Fixed ipv6 icmp
Need ip6tables to accept ipv6-icmp not icmp.

The ip6tables command accepts the argument icmp (because it is a valid protocol in /etc/protocols), but the icmp protocol does not make sense in a ipv6 chain. You want ipv6-icmp instead!

This could lead to some hard to find bugs in dual stacked networks.